backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
sentence-transformers>=3.0.0
torch>=2.1.0
transformers>=4.36.0
groq>=0.4.0
numpy>=1.24.0
